+++ gamedata/shaders/r2/mblur.h
@@ -8,7 +8,7 @@ float3 	mblur	(float2 UV, float3 pos, float3 c_original)	{ return c_original; }
 
 uniform float4x4	m_current;
 uniform float4x4	m_previous;
-uniform half2 	m_blur;		// scale_x / 12, scale_y / 12
+uniform float2 	m_blur;		// scale_x / 12, scale_y / 12
 
 #define MBLUR_SAMPLES 	half(12)
 #define MBLUR_CLAMP	half(0.001)
@@ -18,7 +18,7 @@ float3 	mblur	(float2 UV, float3 pos, float3 c_original)	{
 
 	float4 	p_current	= mul	(m_current,	pos4);
 	float4 	p_previous 	= mul	(m_previous,	pos4);
-	half2 	p_velocity 	= m_blur * ( (p_current.xy/p_current.w)-(p_previous.xy/p_previous.w) );
+	float2 	p_velocity 	= m_blur * ( (p_current.xy/p_current.w)-(p_previous.xy/p_previous.w) );
 		p_velocity	= clamp	(p_velocity,-MBLUR_CLAMP,+MBLUR_CLAMP);
 
 	// For each sample, sum up each sample's color in "Blurred" and then divide
