+++ gamedata/shaders/r2/shadow.h
@@ -28,7 +28,7 @@ half 	sample_sw	(float2 tc, float2 shift, float depth_cmp)
 		depth_cmp-tex2D	(s_smap, Tex11).x);
 	float4 	compare = step	(depth, 0.0f);
 	float2 	fr 		= frac	(Tex00*texsize);
-	half2 	ifr 	= half2	(1,1) - fr;
+	float2 	ifr 	= float2	(1.0f,1.0f) - fr;
 	float4 	fr4 	= float4(ifr.x*ifr.y, ifr.x*fr.y, fr.x*ifr.y,  fr.x*fr.y);
 	return	dot		(compare, fr4);
 }
@@ -88,7 +88,7 @@ half  	sample_hw_f4	(float4 tc,float4 shift){
 
 	float  	texsize 		= 	SMAP_size	;
 	float2 	fr 				= 	frac		(T4.xy * texsize);
-	half2 	ifr 			= 	half2		(1,1) - fr;
+	float2 	ifr 			= 	float2		(1.0f,1.0f) - fr;
 	float4 	fr4 			= 	float4		(ifr.x*ifr.y, ifr.x*fr.y, fr.x*ifr.y,  fr.x*fr.y);
 	float4 	fr4s		 	= 	fr4.zywx	;
 
@@ -115,7 +115,7 @@ uniform sampler2D	jitter1;
 uniform sampler2D	jitter2;
 uniform sampler2D	jitter3;
 uniform float4 		jitterS;
-float4 	test 		(float4 tc, half2 offset)
+float4 	test 		(float4 tc, float2 offset)
 {
 	float4	tcx	= float4 (tc.xy + tc.w*offset, tc.zw);
 	return 	tex2Dproj (s_smap,tcx);
@@ -150,10 +150,10 @@ half 	shadowtest_sun 	(float4 tc, float4 tcJ)			// jittered sampling
 	//float4	J1 	= tex2D	(jitter1,tc_J)*scale;
 
 	const float k = .5f/float(SMAP_size);
-	r.x 	= test 	(tc, J0.xy+half2(-k,-k)).x;
-	r.y 	= test 	(tc, J0.wz+half2( k,-k)).y;
-	r.z		= test	(tc,-J0.xy+half2(-k, k)).z;
-	r.w		= test	(tc,-J0.wz+half2( k, k)).x;
+	r.x 	= test 	(tc, J0.xy+float2(-k,-k)).x;
+	r.y 	= test 	(tc, J0.wz+float2( k,-k)).y;
+	r.z		= test	(tc,-J0.xy+float2(-k, k)).z;
+	r.w		= test	(tc,-J0.wz+float2( k, k)).x;
 
 	return	dot(r,1.h/4.h);
 }
@@ -168,20 +168,20 @@ half 	shadow_high 	(float4 tc)			// jittered sampling
 
 	const float k = 1.f/float(SMAP_size);
 	float4	r;
-	r.x 	= test 	(tc,J0.xy+half2(-k,-k)).x;
-	r.y 	= test 	(tc,J0.wz+half2( k,-k)).y;
+	r.x 	= test 	(tc,J0.xy+float2(-k,-k)).x;
+	r.y 	= test 	(tc,J0.wz+float2( k,-k)).y;
 	
- 	r.z		= test	(tc,J0.xy+half2(-k, k)).z;
- 	r.w		= test	(tc,J0.wz+half2( k, k)).x;
+ 	r.z		= test	(tc,J0.xy+float2(-k, k)).z;
+ 	r.w		= test	(tc,J0.wz+float2( k, k)).x;
 	
 
 	const float k1 = 1.3f/float(SMAP_size);
 	float4	r1;
-	r1.x 	= test 	(tc,-J0.xy+half2(-k1,0)).x;
-	r1.y 	= test 	(tc,-J0.wz+half2( 0,-k1)).y;
+	r1.x 	= test 	(tc,-J0.xy+float2(-k1,0.0f)).x;
+	r1.y 	= test 	(tc,-J0.wz+float2( 0.0f,-k1)).y;
 
-	r1.z	= test	(tc,-2*J0.xy+half2( k1, 0)).z;
- 	r1.w	= test	(tc,-2*J0.wz+half2( 0, k1)).x;
+	r1.z	= test	(tc,-2*J0.xy+float2( k1, 0.0f)).z;
+ 	r1.w	= test	(tc,-2*J0.wz+float2( 0.0f, k1)).x;
 
 	return ( r.x + r.y + r.z + r.w + r1.x + r1.y + r1.z + r1.w )*1.h/8.h;
 }
